refactoring playwright code generation and its working
