+++ .gitignore
@@ -4,3 +4,4 @@ __pycache__
 .cache
 logs
 output
+playwright_tests
