--- playwright-generation/orchestrator/__init__.py
@@ -1,21 +1,0 @@
-"""
-Orchestrator package for AI-enhanced Playwright test generation.
-"""
-
-from orchestrator.agent_orchestrator import PlaywrightAgentOrchestrator
-from orchestrator.agent_specialized import generate_with_specialized_agents
-from orchestrator.extraction_utils import (
-    extract_page_objects_from_coordinator,
-    extract_test_script_from_chat,
-    extract_page_objects_from_specialized,
-    extract_test_script_from_specialized
-)
-
-__all__ = [
-    'PlaywrightAgentOrchestrator',
-    'generate_with_specialized_agents',
-    'extract_page_objects_from_coordinator',
-    'extract_test_script_from_chat',
-    'extract_page_objects_from_specialized',
-    'extract_test_script_from_specialized'
-]
