+++ playwright_generation/agents/__init__.py
@@ -1,0 +1,1 @@
+# agents package
