+++ playwright_generation/generators/__init__.py
@@ -1,0 +1,1 @@
+# generators package
