+++ playwright_generation/prompts/__init__.py
@@ -1,0 +1,1 @@
+# prompts package
